requirements.txt
Django==1.11.7
asgiref==1.1.2
asgi-redis==1.4.3
channels==1.1.7
Celery==3.1.20
django-celery==3.2.2
celery-with-redis
djangorestframework==3.5.3
djangorestframework-jwt==1.11.0
paramiko==2.4.1
Pillow==5.0.0
python-magic==0.4.15
django-storages==1.6.5
PyNaCl==1.2.1
pycparser==2.18
pyasn1==0.4.2
ansible==2.3.3
supervisor
mysql-replication
pymysql
redis
MySQL-python
DBUtils
xlrd
chardet
requests
xlwt
kubernetes-py
jenkinsapi
python-jenkins